refactor(domain,repo,service): 식별자 타입 int → Long 일괄 전환 (orderId-guestId)
- InMemoryData 스키마(Long)와 호환을 위해 Order.id, Order.guestId, Guest.id 등을 Long으로 변경
- 레포지토리/맵 키, 시퀀스 생성기 및 메서드 시그니처·반환타입 전부 Long 기반으로 수정

+++ src/main/java/org/BackOffice/services/loyalty/domain/Guest.java
@@ -1,22 +1,22 @@
 package org.BackOffice.services.loyalty.domain;
 
 public class Guest {
-    private Integer id;
-    private Integer openOrderId;
+    private Long id;
+    private Long openOrderId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public void setOpenOrderId(Integer openOrderId) {
+    public void setOpenOrderId(Long openOrderId) {
         this.openOrderId = openOrderId;
     }
 
-    public Integer getOpenOrderId() {
+    public Long getOpenOrderId() {
         return openOrderId;
     }
